Handle 8bit login
changed the fansi encoding to also trigger 8bits client detection on login. This is a 2 step process as Muddler will wait for the client detection string and then a 'connect' string afterwards to trigger a login instead of just sending as soon as Muddler connects. If you send the strings to soon, you don't get 8bit's alternative fansi login page.